gateway-api/requirements.migrator.txt
aerich==0.7.2
python-dotenv==1.0.0
pypika-tortoise==0.1.6
tortoise==0.1.1
tortoise-orm==0.20.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
yarl==1.9.2
asyncpg==0.29.0
